made changes to the navbar structure
