+++ src/layouts/base_page.js
@@ -7,9 +7,6 @@ import MobileNavBar from '../components/mobile/mobile_navbar';
 function BasePage({children}){
     return(
         <div className="[ base-page ][ flex flex-col ][ bg-[#f7f7f7] ][ h-full overscroll-auto ][ w-screen ]">
-            <MobileTopNavBar/>
-            <MobileNavBar/>
-            <TopNavBar />
             <NavBar />
                 {children}
             <Footer />
